% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Compute of a t-test using permutations}
\usage{
t_test(x, y, mu, var_equal, n_perms)
}
\arguments{
\item{x}{a sample of length nx following a specific distribution}

\item{y}{a sample of length ny following a specific distribution}

\item{var_equal}{a logical scalar, are we considering the variances to be equal or not ?}

\item{n_perms}{the number of permutations}
}
\value{
the pvalue, the test's statistic and the distribution
}
\description{
Compute of a t-test using permutations
}
\examples{
t_test(x,y,mu=0,var_equal=TRUE,10000)
}
