% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_test.R
\name{tf_test}
\alias{tf_test}
\title{computation of a t-test and a var-test simultaneously using permutations}
\usage{
tf_test(x, y, mu, sigma, n_perms)
}
\arguments{
\item{x}{a sample of length nx following a specific distribution}

\item{y}{a sample of length ny following a specific distribution}

\item{n_perms}{the number of permutations}
}
\value{
the pvalue, the test's statistic and the distribution
}
\description{
computation of a t-test and a var-test simultaneously using permutations
}
\examples{
tf_test(x,y,mu=0, sigma=1,10000)
}
