% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_test.R
\name{tf_test}
\alias{tf_test}
\title{Compute a t.test and a var.test simultaneously using permutations}
\usage{
tf_test(x, y, alter, mu, sigma, conf_level, n_perms)
}
\arguments{
\item{x}{a sample of length nx following a specific distribution.}

\item{y}{a sample of length ny following a specific distribution.}

\item{alter}{the type of test computed either "right_tail", "left_tail" or two_tail".}

\item{mu}{a number indicating the difference in means for a two-sample t.test.}

\item{sigma}{a number indicating the ratio between the variances for a two-sample var.test.}

\item{conf_level}{the confidence level of the interval.}

\item{n_perms}{the number of permutations.}
}
\value{
the test's statistic, its p-value,, the estimated difference in means, the estimated ratio in variances, the alternative hypothesis, the method performed, the names of the data
}
\description{
Compute a t.test and a var.test simultaneously using permutations
}
\examples{
tf_test(x,y,alter="two_tail",mu=0,sigma=1,conf_level=0.95,n_perms=10000)
}
