% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_test.R
\name{var_test}
\alias{var_test}
\title{computation of a var-test using permutations}
\usage{
var_test(x, y, sigma, n_perms)
}
\arguments{
\item{x}{a sample of length nx following a specific distribution}

\item{y}{a sample of length ny following a specific distribution}

\item{n_perms}{the number of permutations}
}
\value{
the pvalue, the test's statistic and the distribution
}
\description{
computation of a var-test using permutations
}
\examples{
var_test(x,y,sigma=1,10000)
}
